pyproject.toml
[project]
name = "keyword-scraper"
version = "0.1.0"
description = "Unified trending keyword scraper — Google Trends + Trends24, with taxonomy classification."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.4.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pandas>=2.0",
    "nest-asyncio>=1.6.0",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "requests>=2.28",
    "pytest>=8.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "streamlit>=1.0",
    "httpx>=0.25",
    "python-dotenv>=1.0",
]

[project.scripts]
scrape = "main:main"
